Refactor inventory modules for improved orchestration and separation of concerns
- Introduced `InventoryModule` for single-source inventory operations, which is used by the `WorkItemsModule`.
- Refactored `InventoryDiscoveryModule` to handle multi-org inventory discovery jobs, delegating orchestration to `InventoryOrchestrator`.
- Updated `ModuleServiceCollectionExtensions` to register both `InventoryModule` and `InventoryDiscoveryModule`.
- Modified `WorkItemsModule` to depend on `InventoryModule` instead of `InventoryDiscoveryModule`.
- Adjusted `JobAgentWorker` to route jobs correctly based on job kind, ensuring proper handling of inventory jobs.
