+++ src/DevOpsMigrationPlatform.Infrastructure.Agent/Modules/WorkItemsModule.cs
@@ -42,7 +42,7 @@ public sealed class WorkItemsModule : IModule
     public string Name => "WorkItems";
     public IReadOnlyList<ModuleDependency> DependsOn => new[]
     {
-        new ModuleDependency(typeof(InventoryDiscoveryModule), DependencyPhase.Both),
+        new ModuleDependency(typeof(InventoryModule), DependencyPhase.Both),
         new ModuleDependency(typeof(IdentitiesModule), DependencyPhase.Import),
         new ModuleDependency(typeof(NodesModule), DependencyPhase.Import)
     };
